requirements.txt
Django== 5.1.2 
gunicorn #install gunicorn
python-decouple
psycopg[binary]
dj-database-url
requests
django-storages[s3]
boto3
whitenoise==6.7.0
django-allauth[socialaccount]
django-allauth-ui
django-widget-tweaks
slippers
stripe